pyproject.toml
[project]
name = "chris-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "httptools>=0.6.4",
    "httpx>=0.28.1",
    "pydantic>=2.11.5",
    "pydantic-settings>=2.9.1",
    "python-keycloak>=5.5.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.2",
    "pyjwt[crypto]>=2.10.0",
    "python-jose[cryptography]>=3.3.0",
    "sqlmodel>=0.0.24",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "requests-oauthlib>=2.0.0",
]

[dependency-groups]
dev = [
    "isort>=6.0.1",
    "pytest>=8.3.5",
    "ruff>=0.11.12",
    "mypy>=1.17.1",
    "types-requests>=2.32.4.20250809",
]

[tool.mypy]
python_version = "3.12"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = false
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "keycloak.*",
    "python_keycloak.*",
    "jose.*",
    "requests_oauthlib.*",
]
ignore_missing_imports = true
